% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{create_stattest_function_skeleton}
\alias{create_stattest_function_skeleton}
\title{Template for user-defined statistical test}
\usage{
create_stattest_function_skeleton()
}
\value{
returns function template
}
\description{
Use this template to define customized statistical tests for a specific variable class.
}
\examples{
\dontrun{
create_stattest_function_skeleton()
}
}
