% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{get_vcps}
\alias{get_vcps}
\title{Get valid cluster pairs}
\usage{
get_vcps(sg)
}
\arguments{
\item{sg}{\code{sgi.object} created by \code{sgi::sgi()}}
}
\value{
Returns a matrix where each column is a valid cluster pair and each row corresponds to a sample
}
\description{
Get valid cluster pairs
}
\examples{
\dontrun{
# for a given hclust object 'hc' and outcome data frame 'outcomes'
sg = sgi_init(hc, minsize = 18, outcomes)
get_vcps(sg)
}
}
