% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{ndist}
\alias{ndist}
\title{Combine multiple distance matrices into one for integrated hierarchical clustering}
\usage{
ndist(dlist, w = rep(1, length(dlist)))
}
\arguments{
\item{dlist}{List of distance matrices of class \code{dist}}

\item{w}{Weights for distance matrices in \code{dlist}, if weighted combination is desired; all matrices have equal weight by default}
}
\value{
Combined distance matrix
}
\description{
Combine multiple distance matrices into one for integrated hierarchical clustering
}
\examples{
\dontrun{
d1 = dist(rnorm(10))
d2 = dist(rnorm(10))
d3 = dist(rnorm(10))

# list of distance matrices
dlist = list(d1,d2,d3)
# function call
ndist(dlist)
}

}
\author{
mubu
}
