% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_association_object.R
\name{plot.sgi.assoc}
\alias{plot.sgi.assoc}
\title{Plot subgroup identification results on a tree}
\usage{
\method{plot}{sgi.assoc}(
  as,
  padj_th = 0.05,
  outcomes_text.size = 3.7,
  outcomes_at_middle = TRUE,
  branching_point.size = 3,
  branching_point.color = "white",
  tree_opt = "full_rect",
  plot_only_valid_cluster_pairs = FALSE,
  color.subgroups = NULL,
  color.valid_cluster_pairs = "gray",
  color.invalid_clusters = "gray",
  color_only_sgs_branches = TRUE,
  lwds = c(0.5, 3),
  prune_height = NULL,
  tree_pruning = 4,
  lty.invalid_clusters = 2,
  lty.pruned = 2,
  overlay = TRUE,
  overlay_only_sgs = TRUE,
  overlay.fill = NULL,
  overlay.alpha = 0.085,
  overlay.color = "white",
  overlay.lwd = 0.1,
  overlay.lty = 1,
  draw_cluster_ids = TRUE,
  cluster_ids_for_only_sgs = TRUE,
  cluster_ids.color = "black",
  cluster_ids_text.size = 4,
  cluster_ids_as_label = FALSE,
  vcs.lineend = "round",
  ...
)
}
\arguments{
\item{as}{Association object returned by \code{sgi::sgi_run()}}

\item{padj_th}{Adjusted p-value threshold to filtering out significant enrichments}

\item{outcomes_text.size}{Font size of outcomes on tree}

\item{outcomes_at_middle}{Position of outcomes on the tree, TRUE(default) for middle of samples or FALSE for at branching point on the tree}

\item{branching_point.size}{Size of branching point marking for significant outcomes}

\item{branching_point.color}{Color of branching point marking for significant outcomes}

\item{tree_opt}{Shape of the tree branches: \cr
\code{"full_rect"} Rectangles (default) \cr
\code{"up_tri_down_rect"} Valid cluster pairs as triangles, rest as rectangles \cr
\code{"up_rect_down_tri"} Valid cluster pairs as rectangles, rest as triangles \cr
\code{"full_tri"} Triangles}

\item{plot_only_valid_cluster_pairs}{Plot only valid cluster pairs \code{FALSE} (default)}

\item{color.subgroups}{Single color for all subgroups, if \code{NULL} (default), each subgroup is colored differently}

\item{color.valid_cluster_pairs}{Color of valid cluster pairs, default is black}

\item{color.invalid_clusters}{Color of invalid clusters, default is gray}

\item{color_only_sgs_branches}{Whether only significant subgroups are colored differently, \code{TRUE} (default), otherwise all valid clusters will get different colors.}

\item{lwds}{Line widths, vector of length two containing widths of invalid cluster lines and valid cluster lines}

\item{prune_height}{Pruning height if \code{tree_pruning==1}}

\item{tree_pruning}{Options for tree pruning (i.e. removing bottom branches in the visualization): \cr
\code{0}: no pruning \cr
\code{1}: prune with \code{prune_height} (separate function argument) \cr
\code{2}: adaptive pruning, cut underneath lowest valid cluster pair  \cr
\code{3}: adaptive pruning, cut at highest leaf branching point (cuts away long leaf branches), pruned part is left blank \cr
\code{4}: same as option \code{3}, but pruned part is cut out.}

\item{lty.invalid_clusters}{Line type for invalid clusters, see linetypes options in R}

\item{lty.pruned}{Line type of cut branches after pruning (indicating that pruning has taken place)}

\item{overlay}{Whether valid cluster pairs or subgroups are overlayed with rectangles, \code{TRUE} (default)}

\item{overlay_only_sgs}{\code{TRUE} (default): Overlay subgroups with rectangles,\code{FALSE}: overall all valid cluster pairs with rectangles}

\item{overlay.fill}{Single fill color of overlay rectangles for subgroups. \code{NULL} (default) to fill each subgroups differently.}

\item{overlay.alpha}{Alpha of overlay rectangles}

\item{overlay.color}{Border color of overlay rectangles}

\item{overlay.lwd}{Line width of overlay rectangles}

\item{overlay.lty}{Line type of overlay rectangles}

\item{draw_cluster_ids}{Whether tree is annotated with cluster IDs, \code{TRUE} {default}}

\item{cluster_ids_for_only_sgs}{\code{TRUE} {default}: Cluster IDs are only printed for subgroups, \code{FALSE}: cluster IDs are printed for all valid cluster pairs}

\item{cluster_ids.color}{Color of cluster ID text}

\item{cluster_ids_text.size}{Size of cluster ID text}

\item{cluster_ids_as_label}{Should IDs be printed with \code{geom_label} (\code{TRUE}), or \code{geom_text} (\code{FALSE}, default)}

\item{vcs.lineend}{Line end type of segments of the valid cluster pairs, default is \code{"round"}, options: \code{c("round", "butt", "square")}}
}
\value{
ggplot object of annotated tree
}
\description{
Plots the results of a subgroup identification run from \code{sgi::sgi_run()}.
Offers various options for association highlighting and tailored visualizations.
}
\examples{
\dontrun{

# initialize sgi
sg = sgi::sgi_init(hc, minsize, outcomes)
# run sgi
as = sgi::sgi_run(sg)

# plot sgi results with default options
gg = plot(as)
gg
}
}
\author{
mubu
}
