% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes.R
\name{plot_outcomes}
\alias{plot_outcomes}
\title{Outcome plots for subgroups}
\usage{
plot_outcomes(
  sg,
  as,
  outcome_names = NULL,
  cluster_pairs = NULL,
  padj_th = 0.05,
  drop_nulls = T,
  ...
)
}
\arguments{
\item{sg}{SGI object created by \code{sgi::sgi()}}

\item{as}{SGI association object created by \code{sgi::sgi_run()}}

\item{outcome_names}{Names of outcomes to plot}

\item{cluster_pairs}{Names of cluster pairs to plot, e.g. \code{c("2vs3","4vs5")}}

\item{padj_th}{Adjusted p-value threshold}

\item{drop_nulls}{Whether clinical outcomes with no associations will be excluded from the result list (\code{TRUE}) or included as a \code{NULL} (\code{FALSE})}

\item{...}{}
}
\value{

}
\description{
Generates plots that visualize the differences of a list of outcomes with their associating cluster splits.
}
\examples{
\dontrun{
# initialize SGI structure, minsize is set to 0.05 of total sample size
sg = sgi_init(hc, minsize = 18, outcomes = df_clinicalvars)
# run SGI
as = sgi_run(sg)
# inspect distributions of outcomes in subgroups
ggs = plot_outcomes(sg, as, padj_th = 0.05)
}

}
