% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overview.R
\name{plot_overview}
\alias{plot_overview}
\title{Overview plot}
\usage{
plot_overview(
  gg_tree,
  as,
  outcomes = NULL,
  xdata = NULL,
  h_tree = 25,
  h_xdata = 10,
  padj_th = NULL,
  outcomes_nmax = 10,
  ncut_tree = 0,
  split_markers_for_subgroups = TRUE,
  split_markers_only_for_cuts = FALSE,
  split_marker.size = 2.5,
  split_marker.color = "red",
  split_marker.fill = "red",
  split_marker.alpha = 1,
  split_marker.shape = 19,
  data.color = colorRampPalette(c("blue", "floralwhite", "red"))(100),
  data_cluster_cols = 2,
  data_title = "data",
  draw_legends = F,
  ...
)
}
\arguments{
\item{gg_tree}{ggplot object created by \code{plot(association_object)} (see \code{plot.sgi.assoc()})}

\item{as}{SGI association results, an object of class \code{sgi.assoc} created by \code{sgi::sgi_run()}}

\item{outcomes}{Names of the outcomes to be plotted in heatmap underneath the tree}

\item{xdata}{Data matrix of hierarchical clustering to be plotted underneath the tree (usually the original matrix used for clustering).}

\item{h_tree}{Height of the tree in the plot, relative to 1 plotted outcome strip}

\item{h_xdata}{Height of the data matrix, relative to 1 plotted outcome strip}

\item{padj_th}{Adjusted p-value threshold to decide which splits to mark; if \code{NULL}, set to \code{padj_th} with which \code{gg_tree} was created}

\item{outcomes_nmax}{Maximum number of outcomes, i.e. strip tiles, to be plotted, default is first 10 variables in outcome data.frame.}

\item{ncut_tree}{Cutting the tree to visually emphasize some clusters i.e. number of clusters to highlight \code{ncut_tree>1}}

\item{split_markers_for_subgroups}{Should split markers for subgroups be added into the strip tiles? \code{TRUE} by default}

\item{split_markers_only_for_cuts}{Mark splits only at the level of \code{ncut_tree} if given. FALSE by default.}

\item{split_marker.size}{Size of split marker}

\item{split_marker.color}{Color of split marker}

\item{split_marker.fill}{Filling color of split marker}

\item{split_marker.alpha}{Alpha of split marker}

\item{split_marker.shape}{Shape of the split marker point, one of \code{pch} options}

\item{data.color}{Color palette of data matrix heatmap, see color parameter of \code{pheatmap::pheatmap()}}

\item{data_cluster_cols}{Should data matrix columns be clustered? \code{0}: data not clustered, \code{1} data clustered, but cluster tree not shown, \code{2} data clustered and tree shown}

\item{data_title}{Title to be added to the right of data matrix}

\item{draw_legends}{Draw legends? \code{FALSE} by default; Note: Legends are returned as a list of \code{grid} objects within a ggplot object}

\item{...}{}
}
\value{
Overview plot of SGI analysis
}
\description{
Plots all parts of an SGI analysis in one combined overview plot. Shows the annotated SGI association tree on top, and outcomes and data matrix heatmaps underneath.
}
\examples{
\dontrun{

# hierarchical clustering
hc = hclust(data_matrix)
# SGI initialization
sg = sgi::sgi_init(hc, minsize, outcomes)
# association analysis
as = sgi::sgi_run(sg)
# plot the association analysis tree
gg_tree = plot(as, padj_th = 0.001)
# plot overview
plot_overview( gg_tree, as, outcomes, data_matrix, draw_legends = T, ncut_tree = 3)

}

}
\author{
mubu
}
