% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{print.sgi.assoc}
\alias{print.sgi.assoc}
\title{Print subgroup identification analysis results}
\usage{
\method{print}{sgi.assoc}(as, padj_th = 0.05, by_clusters = T, ...)
}
\arguments{
\item{as}{\code{sgi.assoc} object created by \code{sgi::sgi_run()}}

\item{padj_th}{Adjusted p-value threshold for filtering results}

\item{by_clusters}{\code{TRUE}: Summarize results at by clusters; \code{FALSE}: Summarize results by outcomes}
}
\description{
Print results contained in an \code{sgi.assoc} object created by \code{sgi::sgi_run()}
}
\examples{
\dontrun{
sg = sgi::sgi_init(hc, minsize, outcomes)
as = sgi::sgi_run(sg)
print(as, padj_th = 0.01)
}
}
