% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgi_object.R
\name{print.sgi.object}
\alias{print.sgi.object}
\title{Print an \code{sgi.object}}
\usage{
\method{print}{sgi.object}(sg, summary = F, outcomes = T)
}
\arguments{
\item{sg}{\code{sgi.object} created by \code{sgi::sgi_init()}}

\item{summary}{Print summary of valid clusters? \code{FALSE} by default}

\item{outcomes}{Print outcomes? \code{TRUE} by default}
}
\description{
Print function for \code{sgi.object} created by \code{sgi::sgi_init()}
}
\examples{
\dontrun{
sg = sgi::sgi_init(hc, minsize, outcomes)
sg
print(sg, summary = T, outcomes = F)
}

}
\author{
mubu
}
