% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgi_object.R
\name{sgi_init}
\alias{sgi_init}
\title{Initialize subgroup identification analysis}
\usage{
sgi_init(hc, minsize, outcomes, user_defined_tests = c())
}
\arguments{
\item{hc}{Hierarchical clustering object of class \code{hclust}}

\item{minsize}{Minimum size for a cluster to be considered}

\item{outcomes}{\code{data.frame} of outcomes to be tested; if this data frame has different rownames than \code{hc$labels}, a warning will be thrown}

\item{user_defined_tests}{\link{optional} Named list of functions for user-defined tests. The keys of the list are the classes for which the functions are defined, the values are the functions. See SGI tutorial for more details.}
}
\value{
Object of class \code{sgi.object} which contains all required information to run subgroup identification.
}
\description{
This is a constructor function which initializes the subgroup identification process
}
\examples{
\dontrun{
sg = sgi::sgi_init(hc, minsize, outcomes)
sg
}

}
\author{
mubu
}
