% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{sgi_run}
\alias{sgi_run}
\title{Subgroup Identification Analysis}
\usage{
sgi_run(sg)
}
\arguments{
\item{sg}{an \code{sgi.object} created by \code{sgi::sgi_init()}}
}
\value{
association results as a list which contains
\item{results}{statistical results correponding to each outcome}
\item{cluster_pairs}{summary of cluster pairs, e.g. cluster ids, sample sizes, tree levels etc.}
\item{hc}{ \code{hclust} object given to \code{sgi} }
\item{minsize}{minimum cluster size threshold}
}
\description{
This function runs the subgroup identification analysis initialized by \code{sgi::sgi_init()} and returns statistics of subgroup enrichment for all valid cluster pairs.
Statistical tests are decided based on the type of outcomes. T-test, Fisher's exact tests/Chi-square tests, and log-rank test are employed for continuous, dichotomous/factor, and survival outcomes, respectively.
}
\examples{
\dontrun{

sg = sgi::sgi_init(hc, minsize, outcomes)
as = sgi::sgi_run(sg)
as
}

}
\author{
mubu
}
