% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{summary.sgi.assoc}
\alias{summary.sgi.assoc}
\title{Summary of subgroup identification analysis}
\usage{
\method{summary}{sgi.assoc}(as, padj_th = 0.05, by_clusters = NULL)
}
\arguments{
\item{as}{\code{sgi.assoc} object created by \code{sgi::sgi_run()}}

\item{padj_th}{Adjusted p-value threshold for filtering results}

\item{by_clusters}{\code{TRUE}: Summarize results at by clusters; \code{FALSE}: Summarize results by outcomes}
}
\value{
Summary of SGI results
}
\description{
Summary function that returns a result summary for a \code{sgi.assoc} object created by \code{sgi::sgi_run()}
}
\examples{
\dontrun{
sg = sgi::sgi_init(hc, minsize, outcomes)
as = sgi::sgi_run(sg)
summary(as, padj_th = 0.01)
}
}
