% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgi_object.R
\name{summary.sgi.object}
\alias{summary.sgi.object}
\title{Summary of \code{sgi.object}}
\usage{
\method{summary}{sgi.object}(sg, summary = T, silent = F)
}
\arguments{
\item{sg}{\code{sgi.object} created by \code{sgi::sgi_init()}}

\item{summary}{Whether unpaired valid clusters should be summarized or not. \code{TRUE} (default)}

\item{silent}{\code{TRUE} returns the object but does not print, default is \code{FALSE}}
}
\value{
summary of valid clusters
}
\description{
Summary function that returns information of valid clusters of \code{sgi.object} created by \code{sgi::sgi_init()}
}
\examples{
\dontrun{
sg = sgi::sgi_init(hc, minsize, outcomes)
summary( sg )
}

}
\author{
mubu
}
